折る方向と枚数を決定するカスタムフック
